refactor - modify products package structure
- 상품과 관련된 패키지 구조를 에그리거트 중심 구조로 변경했습니다.
