+++ src/main/java/kitchenpos/bo/MenuBo.java
@@ -3,10 +3,10 @@ package kitchenpos.bo;
 import kitchenpos.dao.MenuDao;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.dao.MenuProductDao;
-import kitchenpos.dao.ProductDao;
+import kitchenpos.products.dao.ProductDao;
 import kitchenpos.model.Menu;
 import kitchenpos.model.MenuProduct;
-import kitchenpos.model.Product;
+import kitchenpos.products.model.Product;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
