+++ src/main/java/kitchenpos/products/dao/JdbcTemplateProductDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.products.dao;
 
-import kitchenpos.model.Product;
+import kitchenpos.products.model.Product;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
