+++ src/main/java/kitchenpos/products/model/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.model;
+package kitchenpos.products.model;
 
 import java.math.BigDecimal;
