+++ src/test/java/kitchenpos/Fixtures.java
@@ -3,6 +3,7 @@ package kitchenpos;
 import kitchenpos.model.*;
 import kitchenpos.orders.model.Order;
 import kitchenpos.orders.model.OrderStatus;
+import kitchenpos.products.model.Product;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
