+++ src/test/java/kitchenpos/bo/MenuBoTest.java
@@ -3,7 +3,7 @@ package kitchenpos.bo;
 import kitchenpos.dao.MenuDao;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.dao.MenuProductDao;
-import kitchenpos.dao.ProductDao;
+import kitchenpos.products.dao.ProductDao;
 import kitchenpos.model.Menu;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
